First Update
I made a tkinter gui with a model included. Currently has an entry field and label, which update based on mood. I chose the BERT model, which supports multiple languages and is primarily used to sense sentimentality. This model outputs a score out of 1-5 and its confidence in the score as a percentage, and I've used that along with tkinter entry boxes and labels to create a prototype. My next action will be to get the spotify api in order to classify these ratings and get playlists.